requirements.txt
  
pyttsx3  
SpeechRecognition  
wikipedia  
PyAudio  
pipwin  
google-generativeai  
python-dotenv 
